Corrections and empty functions
